Renamed html-errors
Renamed the html-errors folder to errors
