+++ src/app/components/errors/not-found/not-found.component.scss
