+++ src/app/components/errors/server-error/server-error.component.scss
